// src/app/model/Operation.ts
export class Operation{
  caiIdCaisse: string;
  commenatire: string;
  dateOperation: Date;
  idCaisse: string;
  idFournisseur: string;
  idOperation: string;
  idPointVente: string;
  idTypeOperation: string;
  idUtilisateur: string;
  montant: number;
  numCommande: number;
  statut: string;
  idEmploye: string;
}